Queue parallelized COPIES for gpbackup_helper
To reduce the amount of time gpbackup_helper is spent blocked waiting
for data to copy. Gpbackup can now queue COPY commands to speed up
gpbackup if the database has cpu cycles to spare. This is done by
setting the --jobs flag when --single-data-file flag is being used.
This can improve backup times significantly for databases that contain a
lot of tables with very little data where the majority of backup time is
spent waiting for gpdb to initiate COPY.

Co-authored-by: Kevin Yeap <kyeap@vmware.com>
Co-authored-by: Brent Doil <bdoil@vmware.com>